Event delegation
eliminated for loop to create event listeners, as they were re-assigning the listeners to the last element chosen (the last iteration of the for loop). Replaced with a single event listener for clicks on the page. The listener checks if the target of the click was a button, then gets the button id, finds the associated card, and begins toggling classes within both. This works!!!